+++ lib/empathetech_flutter_ui.dart
@@ -9,7 +9,6 @@ library empathetech_flutter_ui;
 // Classes //
 
 // Responsive design
-export 'src/classes/responsive-design/EzRow.dart';
 export 'src/classes/responsive-design/EzRowCol.dart';
 export 'src/classes/responsive-design/EzScrollView.dart';
 export 'src/classes/responsive-design/EzSpacer.dart';
@@ -18,33 +17,28 @@ export 'src/classes/responsive-design/EzTransitions.dart';
 
 // Screen readers
 export 'src/classes/screen-readers/EzImage.dart';
-
-// Screen readers: text
-export 'src/classes/screen-readers/text/EzLink.dart';
-export 'src/classes/screen-readers/text/EzPlainText.dart';
-export 'src/classes/screen-readers/text/EzText.dart';
-export 'src/classes/screen-readers/text/EzTextBlock.dart';
-export 'src/classes/screen-readers/text/EzTextStyle.dart';
-export 'src/classes/screen-readers/text/EzWarning.dart';
-export 'src/classes/screen-readers/text/EzWebLink.dart';
+export 'src/classes/screen-readers/EzLink.dart';
+export 'src/classes/screen-readers/EzText.dart';
+export 'src/classes/screen-readers/EzWarning.dart';
+export 'src/classes/screen-readers/EzWebLink.dart';
 
 // User-customization
 export 'src/classes/user-customization/EzAppProvider.dart';
+export 'src/classes/user-customization/EzColorSetting.dart';
 export 'src/classes/user-customization/EzConfig.dart';
-
-// User-customization: settings
-export 'src/classes/user-customization/settings/EzColorSetting.dart';
-export 'src/classes/user-customization/settings/EzDominantHandSwitch.dart';
-export 'src/classes/user-customization/settings/EzFontSetting.dart';
-export 'src/classes/user-customization/settings/EzImageSetting.dart';
-export 'src/classes/user-customization/settings/EzResetButton.dart';
-export 'src/classes/user-customization/settings/EzSliderSetting.dart';
-export 'src/classes/user-customization/settings/EzThemeModeSwitch.dart';
+export 'src/classes/user-customization/EzDominantHandSwitch.dart';
+export 'src/classes/user-customization/EzFontSetting.dart';
+export 'src/classes/user-customization/EzImageSetting.dart';
+export 'src/classes/user-customization/EzResetButton.dart';
+export 'src/classes/user-customization/EzSliderSetting.dart';
+export 'src/classes/user-customization/EzThemeModeSwitch.dart';
 
 // General
 export 'src/classes/EzAlertDialog.dart';
+export 'src/classes/EzRow.dart';
 export 'src/classes/EzBackAction.dart';
 export 'src/classes/EzScreen.dart';
+export 'src/classes/EzTextStyle.dart';
 export 'src/classes/EzVideoPlayer.dart';
 
 // Constants //
