+++ lib/src/classes/EzTextStyle.dart
