+++ lib/src/classes/screen-readers/EzText.dart
